+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ Alpha (do not use these versions in production)
 * optimized templates
 * added stylesheet
 * added model functions for different calculations
+* added tests
 
 0.0.2 (2016-04-18)
 ------------------